requirements.txt
absl-py==0.13.0
albumentations==1.1.0
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach==4.1.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
Cython==0.29.24
debugpy==1.5.1
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
easydict==1.9
efficientnet==1.1.1
-e git+https://github.com/lukemelas/EfficientNet-PyTorch@7e8b0d312162f335785fb5dcfa1df29a75a1783a#egg=efficientnet_pytorch
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
imageio==2.10.3
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916842152/work
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi @ file:///C:/ci/jedi_1611315820849/work
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core==4.9.1
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets==1.0.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mritopng==2.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
notebook @ file:///C:/ci/notebook_1635393701858/work
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydicom==2.2.2
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pylibjpeg==1.3.0
pylibjpeg-libjpeg==1.2.0
pylibjpeg-openjpeg==1.1.1
pylibjpeg-rle==1.1.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pypng==0.0.21
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy==1.11.2
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.4.1
seaborn==0.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.12.1
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm==4.62.2
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
